+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 All notable changes to the "sqlrunner" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
 
-- Initial release
+### Added
+
+- MySQL connection adapter